App.py login
Created a basic login screen. The current script doesn't include logging into the screen. It future changes add a frame for all of the login screen. I recommend changing the screen so its a function then call inside the init. If possible create a home screen so you can select login or signup and also give more information on the todo app.