Initialize sql
Create a sql folder (Might have to change name later) for my sql scripts. The files create_todo, load, login named accordingly. the load_database file will be used to load the database and create the a connection and cursor object. the objects will be used inside the other files mentioned before as a parameters to the classes.