Database
Added the basic needs for a database. As mentioned before created the load_database file that has the classes and functions to create and load a database. Also added a function that ends the connection to the database. Created a basic login script that  Checks if the username or email is invalid or if the password is invalid.